fix: divide component and main page
