+++ src/components/Header.jsx
@@ -1,0 +1,7 @@
+const Header = () => (
+    <h1 className="header">
+        比賽紀錄
+    </h1>
+);
+
+export default Header;
